==> _notebooks/Participants/04_wave_2_estimate_summary_reliabilities.Rmd <==
---
title: "summary reliabilities"
author: "andrew demetriou"
---

```{r setup, include=FALSE}
library('here')                 # file logistics
library('dplyr')                # data manipulation
library('tidyr')                # data manipulation
library('psych')                # ICC2k
library('patchwork')  # merge multiple plots

values <- c("POWER", "ACHIEVEMENT", "HEDONISM",  
            "STIMULATION", "SELF", "UNIVERSALISM", 
            "BENEVOLENCE", "TRADITION",  
            "CONFORMITY", "SECURITY")
```


```{r}
# load subsample lists
file_path <- here("_data", "_intermediary_data")

wave_2_dfs <- readRDS(here("_data", "_participant_scores", "wave_2", "wave_2_dfs.RDS"))

wave_2_dfs <- lapply(wave_2_dfs, function(df){
  df %>% 
    filter(item_ID != "attention") %>%
    mutate(Ethnicity = case_when(
      Ethnicity == "South East Asian" ~ "Asian",
      Ethnicity == "South Asian" ~ "Asian",
      Ethnicity == "East Asian" ~ "Asian",
      Ethnicity == "CONSENT_REVOKED" ~ NA, 
      .default = Ethnicity))
})

wave_2_df <- wave_2_dfs[[1]] %>% 
  select(item_ID, participant_ID, all_of(values))

```

```{r}
format_ICC_df <- function(df){
  df <- df %>% 
    select(ICC, type, value, `lower bound`, `upper bound`) %>% 
    pivot_wider(values_from = c(ICC, `lower bound`, `upper bound`), 
                names_from = type)
  
  colnames(df) <- c("value", 
                    "ICC2", "ICC2k", 
                    "ICC2_lower", "ICC2k_lower", 
                    "ICC2_upper", "ICC2k_upper")
  
  return(df)
}
```


```{r}
make_overall_icc_df <- function(df){

  # iterate over list of values
  list_of_icc_dfs <- lapply(values,function(value){
    
  # select columns
  df <- df %>% dplyr::select(participant_ID, item_ID, value) %>%
    # pivot so that participants are columns
    pivot_wider(names_from = participant_ID, values_from = value)
  
  # remove item_ID column
  icc_df <- df %>% select(-item_ID) %>% 
    # pass to ICC function
    psych::ICC()
  
  # select relevant ICC output
  icc_df <- icc_df$results %>% filter(type =="ICC2" | type =="ICC2k")
  icc_df$value <- value
  return(icc_df)
})
  
  # bind list of icc dfs
  icc_df <- bind_rows(list_of_icc_dfs)
  
  rownames(icc_df) <- NULL
  
  icc_df <- format_ICC_df(icc_df)
  
  return(icc_df)
}
```

```{r}
icc_df <- make_overall_icc_df(wave_2_df %>% select(item_ID, participant_ID, all_of(values)))
```

```{r}
plot_summary_by_metric <- function(df, metric, metric_lower, metric_upper, text) {

   df %>%
      ggplot(aes(y=value, x={{metric}}, color = value)) +
      geom_point() +
      geom_errorbar(aes(xmin = {{metric_lower}}, xmax = {{metric_upper}})) +
      theme_minimal() +
      xlim(-.5, 1) +
      geom_vline(xintercept = 0.75, color = "red", size = .2 ) +
      scale_color_viridis_d(option = "plasma") +
      theme(legend.position="none") +
      ggtitle(paste0(text))
    
}
```

```{r}
p1 <- plot_summary_by_metric(icc_df, ICC2k, ICC2k_lower, ICC2k_upper, "")
p2 <- plot_summary_by_metric(icc_df, ICC2, ICC2_lower, ICC2_upper, "") +
            theme(axis.text.y = element_blank(),
                axis.title.y = element_blank(),
                axis.ticks.y = element_blank())
 
    p1 + p2 + plot_layout(ncol = 2)
```


```{r}
plot_grid(
  plot_summary_by_metric(icc_df, ICC2k, ICC2k_lower, ICC2k_upper, ""),
  plot_summary_by_metric(icc_df, ICC2, ICC2_lower, ICC2_upper, ""))

ggsave(here("images", "wave_participants_ICC2_ICC2k.png"))
```
